+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [TD-2591] Include the number of related concepts on concept view
+- [TD-3164] Service accounts can search business concepts and view domains
 
 ### Changed